nightmare-plugin: render current attack overlay over overheads
