+++ runelite-client/src/main/java/net/runelite/client/plugins/nightmareofashihama/NightmareOfAshihamaOverlay.java
@@ -1,36 +1,25 @@
 package net.runelite.client.plugins.nightmareofashihama;
 
-import net.runelite.api.Point;
-import net.runelite.api.*;
+import net.runelite.api.Client;
+import net.runelite.api.GraphicsObject;
+import net.runelite.api.NPC;
+import net.runelite.api.Perspective;
 import net.runelite.api.coords.LocalPoint;
-import net.runelite.client.game.SkillIconManager;
 import net.runelite.client.ui.overlay.Overlay;
 import net.runelite.client.ui.overlay.OverlayLayer;
 import net.runelite.client.ui.overlay.OverlayPosition;
 import net.runelite.client.ui.overlay.OverlayUtil;
-import net.runelite.client.util.ImageUtil;
 
 import javax.inject.Inject;
 import java.awt.*;
-import java.awt.image.BufferedImage;
 
 public class NightmareOfAshihamaOverlay extends Overlay
 {
-	private static final Color COLOR_ICON_BACKGROUND = new Color(0, 0, 0, 128);
-	private static final Color COLOR_ICON_BORDER = new Color(0, 0, 0, 255);
-	private static final int ICON_WIDTH = 25;
-	private static final int ICON_HEIGHT = 25;
-	private static final int OVERLAY_ICON_DISTANCE = 30;
-	private static final int OVERLAY_ICON_MARGIN = 12;
-
 	private Client client;
 	private NightmareOfAshihamaPlugin plugin;
 
 	@Inject
 	private NightmareOfAshihamaConfig config;
-
-	@Inject
-	private SkillIconManager iconManager;
 
 	@Inject
 	public NightmareOfAshihamaOverlay(Client client, NightmareOfAshihamaPlugin plugin)
@@ -41,31 +30,12 @@ public class NightmareOfAshihamaOverlay extends Overlay
 		this.plugin = plugin;
 	}
 
-	private BufferedImage getAttackStyleIcon(NightmareOfAshihama.AttackStyle attackStyle)
-	{
-		switch (attackStyle)
-		{
-			case MELEE:
-				return iconManager.getSkillImage(Skill.ATTACK);
-			case RANGED:
-				return ImageUtil.resizeImage(iconManager.getSkillImage(Skill.RANGED), 25, 23);
-			case MAGIC:
-				return iconManager.getSkillImage(Skill.MAGIC);
-		}
-		return null;
-	}
-
 	@Override
 	public Dimension render(Graphics2D graphics)
 	{
 		NightmareOfAshihama nightmare = plugin.getNightmare();
 		if (nightmare != null)
 		{
-			if (config.displayCurrentAttack())
-			{
-				renderCurrentAttackOverhead(graphics, nightmare);
-			}
-
 			if (config.highlightActiveTotems())
 			{
 				renderActiveTotemsHighlights(graphics, nightmare);
@@ -77,52 +47,6 @@ public class NightmareOfAshihamaOverlay extends Overlay
 			}
 		}
 		return null;
-	}
-
-	private void renderCurrentAttackOverhead(Graphics2D graphics, NightmareOfAshihama nightmare)
-	{
-		if (nightmare.getCurrentAttackStyle() == null)
-		{
-			return;
-		}
-
-		LocalPoint localPoint = nightmare.getNpc().getLocalLocation();
-		if (localPoint != null)
-		{
-			net.runelite.api.Point point = Perspective.localToCanvas(client, localPoint, client.getPlane(),
-					nightmare.getNpc().getLogicalHeight() + 16);
-
-			if (point != null)
-			{
-				point = new Point(point.getX(), point.getY());
-				int totalWidth = OVERLAY_ICON_MARGIN;
-				totalWidth += ICON_WIDTH;
-
-				int bgPadding = 4;
-				int currentPosX = 0;
-
-				graphics.setStroke(new BasicStroke(2));
-				graphics.setColor(COLOR_ICON_BACKGROUND);
-				graphics.fillOval(
-						point.getX() - totalWidth / 2 + currentPosX - bgPadding,
-						point.getY() - ICON_HEIGHT / 2 - OVERLAY_ICON_DISTANCE - bgPadding,
-						ICON_WIDTH + bgPadding * 2,
-						ICON_HEIGHT + bgPadding * 2);
-
-				graphics.setColor(COLOR_ICON_BORDER);
-				graphics.drawOval(
-						point.getX() - totalWidth / 2 + currentPosX - bgPadding,
-						point.getY() - ICON_HEIGHT / 2 - OVERLAY_ICON_DISTANCE - bgPadding,
-						ICON_WIDTH + bgPadding * 2,
-						ICON_HEIGHT + bgPadding * 2);
-
-				graphics.drawImage(
-						getAttackStyleIcon(nightmare.getCurrentAttackStyle()),
-						point.getX() - totalWidth / 2 + currentPosX,
-						point.getY() - ICON_HEIGHT / 2 - OVERLAY_ICON_DISTANCE,
-						null);
-			}
-		}
 	}
 
 	private void renderActiveTotemsHighlights(Graphics2D graphics, NightmareOfAshihama nightmare)
